refactor: background video just used within contained layout for now
