+++ libs/journeys/ui/src/components/Card/ContainedCover/ContainedCover.tsx
@@ -9,8 +9,8 @@ import { ImageFields } from '../../Image/__generated__/ImageFields'
 import { VideoFields } from '../../Video/__generated__/VideoFields'
 import { useJourney } from '../../../libs/JourneyProvider'
 import { getJourneyRTL } from '../../../libs/rtl'
-import BackgroundVideo from '../BackgroundVideo'
 import OverlayContent from '../OverlayContent'
+import BackgroundVideo from './BackgroundVideo'
 
 import 'videojs-youtube'
 import 'video.js/dist/video-js.css'
